Rename gpu -> parallelization
