+++ nrpy/helpers/parallelization/cuda_utilities.py
@@ -6,7 +6,7 @@ Authors: Samuel D. Tootle; sdtootle **at** gmail **dot** com
 
 import nrpy.c_function as cfc
 import nrpy.params as par  # NRPy+: Parameter interface
-from nrpy.helpers.gpu.gpu_kernel import GPU_Kernel
+from nrpy.helpers.parallelization.gpu_kernel import GPU_Kernel
 
 """
     Define the default launch dictionary for CUDA kernels.
