+++ nrpy/infrastructures/BHaH/MoLtimestepping/MoL_rk_substep.py
@@ -15,7 +15,7 @@ import sympy as sp
 import nrpy.c_function as cfc
 from nrpy.c_codegen import c_codegen
 from nrpy.helpers.generic import superfast_uniq
-from nrpy.helpers.gpu.utilities import generate_kernel_and_launch_code
+from nrpy.helpers.parallelization.utilities import generate_kernel_and_launch_code
 
 ##############################################################################
 # Parallelization check and set:
